Move LLM message extraction into own object
