+++ src/app/features/editor/fromLLM/utils/llm-conversation-creator.spec.ts
@@ -1,0 +1,7 @@
+import { LlmConversationCreator } from './llm-conversation-creator';
+
+describe('LlmConversationCreator', () => {
+  it('should create an instance', () => {
+    expect(new LlmConversationCreator()).toBeTruthy();
+  });
+});
